remove copy-paste + add TransformationProcess + remove 'client' module
