+++ build.sbt
@@ -66,7 +66,7 @@ lazy val root = Project("modelmatrix", file(".")).
   settings(publish :=()).
   settings(publishLocal :=()).
   settings(unidocSettings: _*).
-  aggregate(modelmatrixCore, modelmatrixCli, modelMatrixClient)
+  aggregate(modelmatrixCore, modelmatrixCli)
 
 
 // Model Matrix projects
@@ -78,7 +78,3 @@ lazy val modelmatrixCli =
   ModelMatrixProject("modelmatrix-cli")
     .dependsOn(modelmatrixCore)
     .settings(flywaySettings: _*)
-
-lazy val modelMatrixClient =
-  ModelMatrixProject("modelmatrix-client")
-    .dependsOn(modelmatrixCore)
