+++ modelmatrix-cli/src/main/scala/com/collective/modelmatrix/cli/featurize/SparseFeaturization.scala
@@ -1,15 +1,14 @@
 package com.collective.modelmatrix.cli.featurize
 
-import com.collective.modelmatrix.catalog.ModelMatrixCatalog
+import com.collective.modelmatrix.ModelMatrix.PostgresModelMatrixCatalog
 import com.collective.modelmatrix.cli.{Source, _}
 import com.collective.modelmatrix.transform.Transformer
-import com.collective.modelmatrix.{Labeling, Featurization, ModelMatrix}
+import com.collective.modelmatrix.{Featurization, Labeling, ModelMatrix}
 import org.apache.spark.mllib.linalg.{DenseVector, SparseVector}
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.types._
 import org.slf4j.LoggerFactory
 
-import scala.concurrent.ExecutionContext
 import scalaz._
 
 case class SparseFeaturization(
@@ -18,7 +17,7 @@ case class SparseFeaturization(
   sink: Sink,
   idColumn: String,
   cacheSource: Boolean
-)(implicit val ec: ExecutionContext @@ ModelMatrixCatalog) extends Script with CliModelCatalog with CliSparkContext {
+) extends Script with PostgresModelMatrixCatalog with CliSparkContext {
 
   private val log = LoggerFactory.getLogger(classOf[ValidateInputData])
