+++ modelmatrix-cli/src/main/scala/com/collective/modelmatrix/cli/instance/ListInstances.scala
@@ -1,16 +1,13 @@
 package com.collective.modelmatrix.cli.instance
 
-import com.collective.modelmatrix.catalog.ModelMatrixCatalog
-import com.collective.modelmatrix.cli.{CliModelCatalog, Script}
+import com.collective.modelmatrix.ModelMatrix.PostgresModelMatrixCatalog
+import com.collective.modelmatrix.cli.Script
 import org.slf4j.LoggerFactory
-
-import scala.concurrent.ExecutionContext
-import scalaz._
 
 
 case class ListInstances(
   modelDefinitionId: Option[Int], name: Option[String]
-)(implicit val ec: ExecutionContext @@ ModelMatrixCatalog) extends Script with CliModelCatalog {
+) extends Script with PostgresModelMatrixCatalog {
 
   private val log = LoggerFactory.getLogger(classOf[ListInstances])
