+++ modelmatrix-cli/src/main/scala/com/collective/modelmatrix/cli/instance/ValidateInputData.scala
@@ -1,20 +1,18 @@
 package com.collective.modelmatrix.cli.instance
 
 import com.collective.modelmatrix.ModelMatrix
-import com.collective.modelmatrix.catalog.ModelMatrixCatalog
+import com.collective.modelmatrix.ModelMatrix.PostgresModelMatrixCatalog
 import com.collective.modelmatrix.cli.{Source, _}
 import com.collective.modelmatrix.transform._
 import org.slf4j.LoggerFactory
 
-import scala.concurrent.ExecutionContext
 import scalaz._
 
 case class ValidateInputData(
   modelDefinitionId: Int,
   source: Source,
   cacheSource: Boolean
-)(implicit val ec: ExecutionContext @@ ModelMatrixCatalog)
-  extends Script with CliModelCatalog with CliSparkContext with Transformers {
+) extends Script with PostgresModelMatrixCatalog with CliSparkContext with Transformers {
 
   private val log = LoggerFactory.getLogger(classOf[ValidateInputData])
