--- modelmatrix-client/build.sbt
@@ -1,3 +1,0 @@
-disablePlugins(sbtassembly.AssemblyPlugin)
-
-libraryDependencies ++= Dependencies.modelmatrixClient
