--- modelmatrix-client/src/main/resources/reference.conf
@@ -1,7 +1,0 @@
-modelmatrix {
-  catalog.db {
-    url = "jdbc:postgresql://localhost/modelmatrix?user=modelmatrix&password=modelmatrix"
-    driver = org.postgresql.Driver
-    keepAliveConnection = true
-  }
-}
